unit tests for ShellCommandConfig, move StreamConfig to java directory
